Add files via upload
This dashboard project in Tableau is the result of applying a range of useful skills including data source connection, creation of worksheets, adding filters, building different visualizations (tables, area charts), formatting, and application of dashboard actions (highlights). The dataset used for this project can be found in the ''Tableau Repository'' you get with the software, or you can contact me directly, to get it for free.